Start to flesh out the story steps using some helper objects to manage the application, auction and auction server.
Because our tests and our app are self-contained but separate, there will
be some duplication of certain values (test auction IDs, credentials)
between the tests and the app.

We could pass these values in to the iPhone app as command-line arguments
but this doesn't reflect how an iPhone app would actually be started so
we will live with the hardcoded value duplication for now.